Moved store to own file
